requirements.txt
python-telegram-bot==21.3
requests==2.32.3
apscheduler==3.10.4
httpx==0.27.0
